+++ gdm/__init__.py
@@ -3,7 +3,7 @@
 import sys
 
 __project__ = 'GDM'
-__version__ = '0.6'
+__version__ = '0.7'
 
 CLI = 'gdm'
 PLUGIN = 'deps'
@@ -12,13 +12,15 @@ DESCRIPTION = "A language-agnostic \"dependency manager\" using Git."
 
 PYTHON_VERSION = 3, 3
 
-if not sys.version_info >= PYTHON_VERSION:  # pragma: no cover (manual test)
+if sys.version_info < PYTHON_VERSION:  # pragma: no cover (manual test)
     exit("Python {}.{}+ is required.".format(*PYTHON_VERSION))
 
 try:
+    # pylint: disable=wrong-import-position
     from .commands import install
     from .commands import update
     from .commands import display as list  # pylint: disable=redefined-builtin
+    from .commands import lock
     from .commands import delete as uninstall
 except ImportError:  # pragma: no cover (manual test)
     pass