+++ gdm/cli.py
@@ -61,9 +61,13 @@ def main(args=None, function=None):
                      help="list of dependencies (`dir` values) to update")
     sub.add_argument('-a', '--all', action='store_true', dest='recurse',
                      help="update all nested dependencies, recursively")
-    sub.add_argument('-L', '--no-lock',
-                     action='store_false', dest='lock', default=True,
-                     help="skip recording of versions for later reinstall")
+    group = sub.add_mutually_exclusive_group()
+    group.add_argument('-l', '--lock',
+                       action='store_true', dest='lock', default=None,
+                       help="enable recording of versions for later reinstall")
+    group.add_argument('-L', '--no-lock',
+                       action='store_false', dest='lock', default=None,
+                       help="disable recording of versions for later reinstall")
 
     # Display parser
     info = "display the current version of each dependency"
@@ -72,6 +76,13 @@ def main(args=None, function=None):
     sub.add_argument('-D', '--no-dirty', action='store_false',
                      dest='allow_dirty',
                      help="fail if a source has uncommitted changes")
+
+    # Lock parser
+    info = "lock the current version of each dependency"
+    sub = subs.add_parser('lock', description=info.capitalize() + '.',
+                          help=info, parents=[debug, project], **shared)
+    sub.add_argument('name', nargs='*',
+                     help="list of dependencies (`dir` values) to lock")
 
     # Uninstall parser
     info = "delete all installed dependencies"
@@ -113,6 +124,9 @@ def _get_command(function, namespace):
         function = commands.display
         kwargs.update(dict(depth=namespace.depth,
                            allow_dirty=namespace.allow_dirty))
+    elif namespace.command == 'lock':
+        function = getattr(commands, namespace.command)
+        args = namespace.name
     elif namespace.command == 'uninstall':
         function = commands.delete
         kwargs.update(force=namespace.force)
@@ -124,7 +138,7 @@ def _get_command(function, namespace):
 def _run_command(function, args, kwargs, exit_msg):
     success = False
     try:
-        log.debug("Running %r command...", function.__name__)
+        log.debug("Running %r command...", getattr(function, '__name__', 'a'))
         success = function(*args, **kwargs)
     except KeyboardInterrupt:
         log.debug("Command canceled")