+++ tests/conftest.py
@@ -1,19 +1,18 @@
-"""pytest configuration."""
+"""Integration tests configuration."""
+# pylint: disable=unused-argument
 
 import os
 
 import yorm
 
-from gdm.test.conftest import pytest_configure  # pylint:disable=unused-import
+from gdm.test.conftest import pytest_configure  # pylint: disable=unused-import
 
 
+# TODO: delete if unused (and files)
 ROOT = os.path.dirname(__file__)
 FILES = os.path.join(ROOT, 'files')
 
 
 def pytest_runtest_setup(item):
-    """pytest setup."""
-    if 'integration' in item.keywords:
-        yorm.settings.fake = False
-    else:
-        yorm.settings.fake = True
+    """Ensure files are created for integration tests."""
+    yorm.settings.fake = False